+++ ic-agent/src/identity/basic.rs
@@ -79,6 +79,7 @@ impl Identity for BasicIdentity {
         Ok(Signature {
             signature: Some(signature.as_ref().to_vec()),
             public_key: self.public_key(),
+            delegations: None,
         })
     }
 }